cleaned and document page added
