+++ FastApiTemplate/.env copy
@@ -1,0 +1,6 @@
+DEBUG=True
+SECRET_KEY=4dde00ea8f92a7d36959
+ALLOWED_HOSTS=127.0.0.1, localhost 
+ALGORITHM=HS256 
+SESSION_EXPIRE_MINUTES=60 
+ACCESS_TOKEN_EXPIRE_MINUTES=10
