+++ FastApiTemplate/depends/Auths/__init__.py
@@ -1,0 +1,2 @@
+from .routes import router
+from .routes import AuthManager
