+++ FastApiTemplate/docs/index.md
@@ -1,0 +1,4 @@
+# FastapiWithAccount
+![](fastapipythn.jpg)
+
+This is a Template repository for FastApi Applications
