--- Templates/pyvenv.cfg
@@ -1,3 +1,0 @@
-home = F:\Program File\python
-include-system-site-packages = false
-version = 3.10.7
